Added numbers

+++ src/string.c
@@ -6,7 +6,7 @@ bool string_eq(String s1, String s2) {
     if (s1.length != s2.length) {
         return false;
     }
-    for (size_t i = 0; i < s1.length; i++) {
+    for (int i = 0; i < s1.length; i++) {
         if (s1.string[i] != s2.string[i])
             return false;
     }
